Move ParticipantGroup
