+++ study/src/org/labkey/study/importer/CreateChildStudyPipelineJob.java
@@ -43,6 +43,7 @@ import org.labkey.api.study.Dataset;
 import org.labkey.api.study.StudySnapshotType;
 import org.labkey.api.study.TimepointType;
 import org.labkey.api.study.Visit;
+import org.labkey.api.study.model.ParticipantGroup;
 import org.labkey.api.view.NotFoundException;
 import org.labkey.api.view.ViewContext;
 import org.labkey.api.writer.MemoryVirtualFile;
@@ -52,7 +53,6 @@ import org.labkey.study.StudySchema;
 import org.labkey.study.model.ChildStudyDefinition;
 import org.labkey.study.model.DatasetDefinition;
 import org.labkey.study.model.ParticipantCategoryImpl;
-import org.labkey.study.model.ParticipantGroup;
 import org.labkey.study.model.ParticipantGroupManager;
 import org.labkey.study.model.ParticipantMapper;
 import org.labkey.study.model.StudyImpl;
