+++ study/src/org/labkey/study/specimen/report/SpecimenVisitReport.java
@@ -27,11 +27,11 @@ import org.labkey.api.specimen.report.SpecimenReportCellData;
 import org.labkey.api.specimen.report.SpecimenReportTitle;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyService;
+import org.labkey.api.study.model.ParticipantGroup;
 import org.labkey.api.util.PageFlowUtil;
 import org.labkey.api.view.ActionURL;
 import org.labkey.api.view.NotFoundException;
 import org.labkey.study.CohortFilter;
-import org.labkey.study.model.ParticipantGroup;
 import org.labkey.study.model.ParticipantGroupManager;
 import org.labkey.study.model.StudyManager;
 import org.labkey.study.model.VisitImpl;
