+++ study/src/org/labkey/study/visitmanager/RelativeDateVisitManager.java
@@ -31,9 +31,9 @@ import org.labkey.api.query.FilteredTable;
 import org.labkey.api.security.User;
 import org.labkey.api.study.DataspaceContainerFilter;
 import org.labkey.api.study.Study;
+import org.labkey.api.study.model.ParticipantGroup;
 import org.labkey.study.CohortFilter;
 import org.labkey.study.StudySchema;
-import org.labkey.study.model.ParticipantGroup;
 import org.labkey.study.model.QCStateSet;
 import org.labkey.study.model.StudyImpl;
 import org.labkey.study.model.StudyManager;
